Moved all s3 tests into one class
to avoid https://github.com/mozilla/OpenWPM/issues/732
Please note that this doesn't close the issue
